Remove timeslot_utils.js file
